+++ packages/ngx-forms/src/lib/datepicker/components/datepicker/datepicker.component.ts
@@ -33,7 +33,7 @@ import {
   DATEPICKER_DATE_MASK,
   DATEPICKER_DEFAULT_ERROR_LABELS,
   DATEPICKER_ERROR_LABELS,
-  DATEPICKER_SEPARATOR_CHAR
+  DATEPICKER_SEPARATOR_CHAR,
 } from '../../datepicker.conf';
 import { DatepickerValidationErrors } from '../../types/datepicker.types';
 import { Interval, IntervalBuilder } from '@acpaas-ui/ngx-utils';
@@ -41,22 +41,23 @@ import { Interval, IntervalBuilder } from '@acpaas-ui/ngx-utils';
 @Component({
   selector: 'aui-datepicker',
   templateUrl: './datepicker.component.html',
-  styleUrls: [
-    './datepicker.component.scss',
-  ],
+  styleUrls: ['./datepicker.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  providers: [{
-    provide: NG_VALUE_ACCESSOR,
-    useExisting: forwardRef(() => DatepickerComponent), // tslint:disable-line:no-forward-ref
-    multi: true,
-  }, {
-    provide: NG_VALIDATORS,
-    useExisting: forwardRef(() => DatepickerComponent), // tslint:disable-line:no-forward-ref
-    multi: true,
-  }],
+  providers: [
+    {
+      provide: NG_VALUE_ACCESSOR,
+      useExisting: forwardRef(() => DatepickerComponent), // eslint-disable-line @angular-eslint/no-forward-ref
+      multi: true,
+    },
+    {
+      provide: NG_VALIDATORS,
+      useExisting: forwardRef(() => DatepickerComponent), // eslint-disable-line @angular-eslint/no-forward-ref
+      multi: true,
+    },
+  ],
 })
 export class DatepickerComponent implements OnInit, OnChanges, OnDestroy, ControlValueAccessor {
-  @ViewChild(FlyoutDirective, {static: true}) flyout: FlyoutDirective;
+  @ViewChild(FlyoutDirective, { static: true }) flyout: FlyoutDirective;
   @Input() id: string;
   @Input() name: string;
   @Input() placeholder = 'dd/mm/yyyy';
@@ -69,9 +70,11 @@ export class DatepickerComponent implements OnInit, OnChanges, OnDestroy, Contro
   @Input() weekdayLabels: WeekdayLabelsConfig;
   @Input() monthLabels: MonthLabelsConfig;
 
+  /* eslint-enable */
   @Output() blur = new EventEmitter<Event>();
+  /* eslint-enable */
 
-  public dateMask = {mask: DATEPICKER_DATE_MASK, showMaskOnHover: false};
+  public dateMask = { mask: DATEPICKER_DATE_MASK, showMaskOnHover: false };
   public formControl: FormControl;
   public selectedDate: Date;
   public isDisabled = false;
@@ -86,39 +89,34 @@ export class DatepickerComponent implements OnInit, OnChanges, OnDestroy, Contro
     public calendarService: CalendarService,
     private formBuilder: FormBuilder,
     private ref: ChangeDetectorRef
-  ) {
-  }
+  ) {}
 
   public ngOnInit(): void {
     this.weekdayLabels = this.weekdayLabels || this.moduleWeekdayLabels;
     this.monthLabels = this.monthLabels || this.moduleMonthLabels;
     this.createInterval();
-    this.formControl = this.formBuilder.control({value: '', disabled: this.isDisabled});
-    this.formControl.valueChanges
-      .pipe(
-        takeUntil(this.componentDestroyed$)
-      )
-      .subscribe((value) => {
-        if (value) {
-          const format = value.split(DATEPICKER_SEPARATOR_CHAR).reverse().join('-');
-          const date = DateHelper.parseDate(format, 'yyyy-MM-dd');
-          if (date) {
-            this.selectedDate = date;
-            this.onChange(date.toISOString());
-          } else {
-            // Change value with original value (and not null or '') so we can add an error in the validate function
-            this.onChange(value);
-          }
+    this.formControl = this.formBuilder.control({ value: '', disabled: this.isDisabled });
+    this.formControl.valueChanges.pipe(takeUntil(this.componentDestroyed$)).subscribe((value) => {
+      if (value) {
+        const format = value.split(DATEPICKER_SEPARATOR_CHAR).reverse().join('-');
+        const date = DateHelper.parseDate(format, 'yyyy-MM-dd');
+        if (date) {
+          this.selectedDate = date;
+          this.onChange(date.toISOString());
         } else {
-          this.selectedDate = null;
-          this.onChange('');
+          // Change value with original value (and not null or '') so we can add an error in the validate function
+          this.onChange(value);
         }
-      });
+      } else {
+        this.selectedDate = null;
+        this.onChange('');
+      }
+    });
   }
 
   public ngOnChanges(changes: SimpleChanges): void {
     if (changes.min || changes.max) {
-       this.createInterval();
+      this.createInterval();
     }
   }
 
@@ -132,15 +130,21 @@ export class DatepickerComponent implements OnInit, OnChanges, OnDestroy, Contro
       return;
     }
     // Create an interval if min/max is filled in
-    this.interval = IntervalBuilder.dateInterval(this.min ? new Date(this.min) : null, this.max ? new Date(this.max) : null)
+    this.interval = IntervalBuilder.dateInterval(
+      this.min ? new Date(this.min) : null,
+      this.max ? new Date(this.max) : null
+    )
       .not()
       .build();
   }
 
   public writeValue(value: string | Date): void {
-    this.selectedDate = typeof value === 'string'
-      ? this.isISODateFormat(value) ? new Date(value) : DateHelper.parseDate(value, 'dd/MM/yyyy')
-      : value;
+    this.selectedDate =
+      typeof value === 'string'
+        ? this.isISODateFormat(value)
+          ? new Date(value)
+          : DateHelper.parseDate(value, 'dd/MM/yyyy')
+        : value;
     const dateString = this.selectedDate ? this.formatDate(this.selectedDate) : '';
     this.formControl.setValue(dateString);
   }
@@ -204,9 +208,11 @@ export class DatepickerComponent implements OnInit, OnChanges, OnDestroy, Contro
     // throw error when out of range
     const range = this.calendarService.getRangeForDate(date, this.range);
 
-    return range.indexOf(date.getDate()) >= 0 ? {
-      range: this.errorLabels.ERRORS_INVALID_RANGE,
-    } : null;
+    return range.indexOf(date.getDate()) >= 0
+      ? {
+          range: this.errorLabels.ERRORS_INVALID_RANGE,
+        }
+      : null;
   }
 
   public handleBlur(e: Event): void {