+++ packages/ngx-layout/src/lib/modal/services/modal.service.spec.ts
@@ -79,7 +79,7 @@ describe('Modal - ModalService', () => {
       'should do nothing if there are no active modals',
       injectService((modalService: ModalService) => {
         const modal = {
-          destroy: () => {}, // tslint:disable-line:no-empty
+          destroy: () => {}, // eslint-disable-line no-empty, @typescript-eslint/no-empty-function
         };
 
         spyOn(modal, 'destroy');