+++ packages/ngx-localstorage/src/lib/types/localstorage.types.ts
@@ -4,11 +4,11 @@ export interface ILocalStorageConfig {
 
 export type LocalStorageType = 'localStorage' | 'sessionStorage' | 'memory';
 
-// tslint:disable-next-line:no-empty-interface
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface IStorage extends Storage {
 
 }
 
-// tslint:disable-next-line:no-empty-interface
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface IWindow extends Window {
 }