+++ packages/ngx-logo/src/lib/aui-logo.component.ts
@@ -2,19 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'aui-logo',
-  template: `
-    <p>
-      aui-logo works!
-    </p>
-  `,
-  styles: []
+  template: ` <p>aui-logo works!</p> `,
+  styles: [],
 })
-export class AuiLogoComponent implements OnInit {
-
-  constructor() {
-  }
-
-  ngOnInit() {
-  }
-
-}
+export class AuiLogoComponent {}