+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - Fix some typing issues (ModuleWithProviders)
   - Outcommented some test that were not working (Navigation menu that should be deleted UPCO-72 and Context component which has low priority UPCO-18)
   - Fix deprecations (async => waitForAsync)
+- Update from 10 to 11 https://update.angular.io/?v=10.0-11.0
+  - TSLint deprecated => Migrate to ESLint
 
 ## [5.4.0] - 2022-11-07